pyproject.toml
[project]
name = "fed-algs"
version = "0.1.0"
description = "A collection of federated learning algorithms"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flwr>=1.18.0",
    "matplotlib>=2.2.5",
    "numpy>=1.12.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=0.21.0",
    "torch>=1.2.0",
    "torchvision>=0.4.0",
]
